requirements.txt
git+https://github.com/jnsougata/discohook
python-Levenshtein
thefuzz[speedup]
requests
uvicorn
PyYAML